requirements.txt
flask==2.3.3
pandas==3.0.2
joblib==1.3.1
scikit-learn==1.8.0
xgboost==2.0.0
tldextract==5.0.1
numpy==2.2.6
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==2.3.7
